refactor: `kana_parser` リネーム
