+++ test/test_kana_converter.py
@@ -2,12 +2,12 @@ from typing import List
 from unittest import TestCase
 
 from voicevox_engine.model import AccentPhrase, Mora, ParseKanaError, ParseKanaErrorCode
-from voicevox_engine.tts_pipeline import kana_parser
-from voicevox_engine.tts_pipeline.kana_parser import create_kana
+from voicevox_engine.tts_pipeline import kana_converter
+from voicevox_engine.tts_pipeline.kana_converter import create_kana
 
 
 def parse_kana(text: str) -> List[AccentPhrase]:
-    accent_phrases = kana_parser.parse_kana(text)
+    accent_phrases = kana_converter.parse_kana(text)
     return accent_phrases
 
 
@@ -59,7 +59,7 @@ class TestParseKana(TestCase):
     def _accent_phrase_marks_base(
         self, text: str, expected_accent_phrases: List[AccentPhrase]
     ) -> None:
-        accent_phrases = kana_parser.parse_kana(text)
+        accent_phrases = kana_converter.parse_kana(text)
         self.assertEqual(expected_accent_phrases, accent_phrases)
 
     def test_accent_phrase_marks(self):
@@ -556,7 +556,7 @@ class TestParseKanaException(TestCase):
         self.assertEqual(err.exception.kwargs, {"position": "2"})
 
         with self.assertRaises(ParseKanaError) as err:
-            kana_parser.parse_kana("ア？ア'")
+            kana_converter.parse_kana("ア？ア'")
         self.assertEqual(
             err.exception.errcode, ParseKanaErrorCode.INTERROGATION_MARK_NOT_AT_END
         )
