+++ test/test_mock_synthesis_engine.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 from voicevox_engine.dev.core import MockCoreWrapper
 from voicevox_engine.dev.synthesis_engine import MockTTSEngine
 from voicevox_engine.model import AccentPhrase, AudioQuery, Mora
-from voicevox_engine.tts_pipeline.kana_parser import create_kana
+from voicevox_engine.tts_pipeline.kana_converter import create_kana
 
 
 class TestMockTTSEngine(TestCase):
