+++ test/test_user_dict_model.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 from pydantic import ValidationError
 
 from voicevox_engine.model import UserDictWord
-from voicevox_engine.tts_pipeline.kana_parser import parse_kana
+from voicevox_engine.tts_pipeline.kana_converter import parse_kana
 
 
 class TestUserDictWords(TestCase):
